Proper map code splitting
