+++ src/map/basemaps/index.js
@@ -1,0 +1,5 @@
+import GL_STYLE from '../glmap/gl-styles/style.json'
+
+const AVAILABLE_BASEMAPS = GL_STYLE.metadata['gfw:basemap-layers']
+
+export default AVAILABLE_BASEMAPS
