+++ src/map/index.js
@@ -1,1 +1,3 @@
-export { default, targetMapVessel, AVAILABLE_BASEMAPS } from './map'
+export { default } from './map'
+export { targetMapVessel } from './store'
+export { default as AVAILABLE_BASEMAPS } from './basemaps'
